`rptest`: move `in_fips_environment()` to `utils/mode_checks`
