+++ tests/rptest/utils/mode_checks.py
@@ -66,3 +66,17 @@ def skip_debug_mode(*args, **kwargs):
         return ignore(args, kwargs)
     else:
         return args[0]
+
+
+def in_fips_environment() -> bool:
+    """
+    Returns True if the file /proc/sys/crypto/fips_enabled is present and
+    contains '1', otherwise returns False.
+    """
+    fips_file = "/proc/sys/crypto/fips_enabled"
+    if os.path.exists(fips_file) and os.path.isfile(fips_file):
+        with open(fips_file, 'r') as f:
+            contents = f.read().strip()
+            return contents == '1'
+
+    return False
